requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
httpx>=0.23.0
jinja2>=3.0.0
python-multipart>=0.0.5
mysql-connector-python>=8.0.0 